[PCLI-124] refactor: remove function  verify link for recovery password of entry class UserController and create new file in dir config and import
